labs/lab3/requirements.txt
langchain>=0.3.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
langchain-chroma>=0.1.0
langchain-huggingface>=0.1.0
chromadb>=0.5.0
openai>=1.40.0
requests>=2.31.0
python-dotenv>=1.0.0
typing-extensions>=4.9.0
sentence-transformers>=3.0.0
